ci: sync `main` to `release`
these changes implement a workflow to automatically sync changes pushed to the
`main` branch to `release`. it does this with a rebase to maintain accurate
commit history.

in the event of a failed rebase (most likely a conflict), the workflow will
raise a PR so that a human can intervene and resolve any conflicts.